Remove pointless casts (#895)
* Remove pointless casts

These are all casts where the value was already of the same type.

* Use var style for zero-value not cast style

+++ scan_rr.go
@@ -1399,7 +1399,7 @@ func setEUI64(h RR_Header, c *zlexer, o, f string) (RR, *ParseError) {
 	if e != nil {
 		return nil, &ParseError{f, "bad EUI68 Address", l}
 	}
-	rr.Address = uint64(i)
+	rr.Address = i
 	return rr, nil
 }
